+++ assets/src/edit-story/app/config/index.js
@@ -1,2 +1,2 @@
 export { default as ConfigProvider } from './configProvider';
-export { default } from './useConfig';
+export { default as useConfig } from './useConfig';
