+++ assets/src/edit-story/app/register_blocks.js
@@ -3,7 +3,7 @@
  */
 import { registerBlockType } from '@wordpress/blocks';
 
-const blocks = require.context( '../../blocks', true, /(?<!test\/)index\.js$/ );
+const blocks = require.context( '../blocks', true, /(?<!test\/)index\.js$/ );
 
 blocks.keys().forEach( ( modulePath ) => {
 	const { name, settings } = blocks( modulePath );
