+++ assets/src/edit-story/app/story/actions/useClearSelection.js
